Migrating `AggregationForm` to TSX, making use of `useFieldTypes`.
